refactor: useLoggedIn 훅을 userInfo 모듈로 이동
